Elimina comentario

+++ app/src/main/java/edu/ucne/composedemo/data/remote/dto/TicketPostMen.kt
@@ -1,7 +1,7 @@
 package edu.ucne.composedemo.data.remote.dto
 
 data class TicketPostMen(
-    val idTicket: Int,
+    val idTicket: Double,
     val enviadoPor: String,
     val mensaje: String,
     val idUsuario: Int
